+++ testdata/singleton_track/sclass_attr_is_not_an_instance_method.rb
@@ -1,0 +1,9 @@
+# The other half of the same fact: the accessor is NOT on instances.
+# MRI raises NoMethodError on the last line.
+class Config
+  class << self
+    attr_accessor :endpoint
+  end
+end
+
+Config.new.endpoint